+++ packages/gcode-renderer-three/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chestnutlabs/gcode-renderer-three
 
+## 0.11.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @chestnutlabs/gcode-colors@0.11.0
+  - @chestnutlabs/toolpath-core@0.11.0
+
 ## 0.10.0
 
 ### Minor Changes